Move Resume Button login within livechatReadOnly file
