--- app/livechat/client/views/app/livechatOnHold.html
@@ -1,14 +1,0 @@
-<template name="livechatOnHold">
-	{{#if roomOpen}}
-		{{#if isPreparing}}
-			{{> loading}}
-		{{else}}
-				<div class="rc-message-box__join">
-					{{{_ "chat_on_hold_due_to_inactivity"}}}
-					<button class="rc-button rc-button--primary rc-button--small rc-message-box__resume-it-button js-resume-it">{{_ "Resume"}}</button>
-				</div>
-		{{/if}}
-	{{else}}
-		<p>{{_ "This_conversation_is_already_closed"}}</p>
-	{{/if}}
-</template>
