+++ app/livechat/client/views/app/livechatReadOnly.html
@@ -4,13 +4,20 @@
 			{{#if isPreparing}}
 				{{> loading}}
 			{{else}}
-				{{#if inquiryOpen}}
+				{{#if isOnHold}}
 					<div class="rc-message-box__join">
-						{{{_ "you_are_in_preview_mode_of_incoming_livechat"}}}
-						<button class="rc-button rc-button--primary rc-button--small rc-message-box__take-it-button js-take-it">{{_ "Take_it"}}</button>
+						{{{_ "chat_on_hold_due_to_inactivity"}}}
+						<button class="rc-button rc-button--primary rc-button--small rc-message-box__resume-it-button js-resume-it">{{_ "Resume"}}</button>
 					</div>
 				{{else}}
-					{{_ "room_is_read_only"}}
+					{{#if inquiryOpen}}
+						<div class="rc-message-box__join">
+							{{{_ "you_are_in_preview_mode_of_incoming_livechat"}}}
+							<button class="rc-button rc-button--primary rc-button--small rc-message-box__take-it-button js-take-it">{{_ "Take_it"}}</button>
+						</div>
+					{{else}}
+						{{_ "room_is_read_only"}}
+					{{/if}}
 				{{/if}}
 			{{/if}}
 		{{/if}}
